+++ components/ShareMarketDetail.tsx
@@ -139,27 +139,29 @@ export default function ShareMarketDetail({
   }
 
   const handleDelete = async () => {
-    try {
-      const response = await fetch(
-        `https://711.ha-ving.store/boards/markets/${id}`,
-        {
-          method: "DELETE",
-          headers: {
-            "Content-Type": "application/json",
-            Authorization:
-              "Bearer eyJhbGciOiJIUzUxMiJ9.eyJpbWFnZSI6Imh0dHBzOi8vYXZhdGFycy5naXRodWJ1c2VyY29udGVudC5jb20vdS83OTI3MDIyOD92PTQiLCJwYXNzd29yZCI6IiQyYSQxMCRleHhmWXAveXZzNHpiY3cyRFNDalZlREFDaTVlcWZma01HaDlsVWwwTXFBRWRUM2h5WDVEeSIsInBob25lIjoiMDEwMTExMTIyMjIiLCJyb2xlcyI6WyJST0xFX1VTRVIiXSwibmlja25hbWUiOiJuaWNrbmFtZTEiLCJpZCI6MjEsInVzZXJuYW1lIjoidXNlciIsImF1dGgiOlt7ImF1dGhvcml0eSI6IlJPTEVfVVNFUiJ9XSwiaWF0IjoxNzE3MDU5MjA5LCJleHAiOjE3MTk2NTEyMDl9.4bpxNGqYITfq2174mngAguJK3gQZ5gl7KzWB8N5eMQ4TV-e8_Ka7xlzCdGH8u6XEoiMywHZwJLM1_7tlAqtt0A"
+    if (confirm("게시글을 삭제하시겠습니까?")) {
+      try {
+        const response = await fetch(
+          `https://711.ha-ving.store/boards/markets/${id}`,
+          {
+            method: "DELETE",
+            headers: {
+              "Content-Type": "application/json",
+              Authorization:
+                "Bearer eyJhbGciOiJIUzUxMiJ9.eyJpbWFnZSI6Imh0dHBzOi8vYXZhdGFycy5naXRodWJ1c2VyY29udGVudC5jb20vdS83OTI3MDIyOD92PTQiLCJwYXNzd29yZCI6IiQyYSQxMCRleHhmWXAveXZzNHpiY3cyRFNDalZlREFDaTVlcWZma01HaDlsVWwwTXFBRWRUM2h5WDVEeSIsInBob25lIjoiMDEwMTExMTIyMjIiLCJyb2xlcyI6WyJST0xFX1VTRVIiXSwibmlja25hbWUiOiJuaWNrbmFtZTEiLCJpZCI6MjEsInVzZXJuYW1lIjoidXNlciIsImF1dGgiOlt7ImF1dGhvcml0eSI6IlJPTEVfVVNFUiJ9XSwiaWF0IjoxNzE3MDU5MjA5LCJleHAiOjE3MTk2NTEyMDl9.4bpxNGqYITfq2174mngAguJK3gQZ5gl7KzWB8N5eMQ4TV-e8_Ka7xlzCdGH8u6XEoiMywHZwJLM1_7tlAqtt0A"
+            }
           }
+        )
+        if (response.ok) {
+          const responseData = await response.json()
+          console.log("삭제 성공:", responseData)
+          router.back()
+        } else {
+          console.error("삭제를 실패했습니다.:", response.statusText)
         }
-      )
-      if (response.ok) {
-        const responseData = await response.json()
-        console.log("삭제 성공:", responseData)
-        router.back()
-      } else {
-        console.error("삭제를 실패했습니다.:", response.statusText)
+      } catch (error) {
+        console.error("에러 발생:", error)
       }
-    } catch (error) {
-      console.error("에러 발생:", error)
     }
   }
 
@@ -210,8 +212,9 @@ export default function ShareMarketDetail({
                 {image_urls && image_urls.length > 0 ? (
                   <Swiper
                     pagination={{
-                      type: "progressbar"
+                      type: "fraction"
                     }}
+                    loop={true}
                     navigation={true}
                     modules={[Pagination, Navigation]}
                     className="w-full">
@@ -257,7 +260,7 @@ export default function ShareMarketDetail({
                       {title}
                     </p>
                   </div>
-                  {price ? (
+                  {category_name === "중고거래" && price ? (
                     <span className="text-grey_900 text-2xl font-semibold">
                       가격:{price}원
                     </span>
